Add support to run script on AWS Lambda
This is a big commit, centered around running dontspendtoomuch.py on
Lambda. One part of it is writing Terraform to define infrastructure.
That terraform file shouldn't live in this repo for long; it should
move into the central scimma/aws-dev repository for a consistent state
management strategy. But for now, here is fine.

Another part of it is work to create Lambda-style build artifacts.
That's all done inside Make. The idea is just that we need to ship up
a zip archive with the script and all of its dependencies.

Finally, there are some little changes to dontspendtoomuch.py. An
entrypoint for lambda invocations is added, and a barebones
send-to-slack function is added. That function doesn't do any
formatting so it's not very useful just yet.

As of this commit, I have tfstate files stored locally. I am not sure
it's worth setting up state management since we'll be moving the .tf
file over to a different repo where that work is already done.